create edit comment file
